backend/requirements.txt
# Cyber Guard Platform - Backend Dependency List
# Purpose: Pin the minimum set of packages for a runnable learning skeleton.
# Inputs/Outputs: Consumed by `pip install -r requirements.txt`.
# Dependencies: FastAPI stack, SQLAlchemy, testing libraries, PDF/report support.
# TODO Checklist:
# - [ ] Review and pin versions approved by your faculty/instructor.
# - [ ] Add linting/formatting/dev tooling (ruff, black, mypy) later.
# - [ ] Add security scanning (pip-audit) in CI.
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.46
alembic==1.14.0
psycopg[binary]==3.2.13
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
httpx==0.27.2
reportlab==4.2.5
pytest==8.3.3
pytest-asyncio==0.24.0
respx==0.21.1
